refactor(api): push student-location resolution and work-time-model validation into services
Move the student-location resolver (StudentLocationSnapshot type, its
ResolveStudentLocation/ResolveStudentLocationWithTime methods, and
ResolveBinaryLocation) from api/common into services/active, where the
underlying attendance/visit/group data originates. api/common re-exports the
types via aliases so existing call sites are unchanged; the batch loader stays
in api/common because its unexported helpers are frozen by an in-package test.

Add work-time-model invariant validation (name, rotation bounds, per-entry
validation, week_index vs rotation_length, duplicate-slot rejection) to
WorkTimeModelService.CreateModel/UpdateModel so the invariant holds for every
caller, not only the HTTP handler. Behavior over the HTTP path is unchanged:
the handler still pre-validates and returns the same 400 responses.

Drop the ResolveStudentLocationWithTime entry from the handler-complexity
ratchet allowlist now that the method lives outside api/.
