+++ backend/services/config/work_time_model_service.go
@@ -2,6 +2,8 @@ package config
 
 import (
 	"context"
+	"errors"
+	"fmt"
 
 	configModels "github.com/moto-nrw/project-phoenix/models/config"
 )
@@ -33,16 +35,50 @@ func (s *WorkTimeModelService) GetModel(ctx context.Context, id int64) (*configM
 // CreateModel persists a new model with its entries and returns the
 // freshly reloaded row.
 func (s *WorkTimeModelService) CreateModel(ctx context.Context, model *configModels.WorkTimeModel, entries []*configModels.WorkTimeModelEntry) (*configModels.WorkTimeModel, error) {
+	if err := validateModelWithEntries(model, entries); err != nil {
+		return nil, err
+	}
 	if err := s.repo.Create(ctx, model, entries); err != nil {
 		return nil, err
 	}
 	return s.repo.FindByID(ctx, model.ID)
 }
 
+// validateModelWithEntries enforces the work-time-model invariants for any
+// caller (HTTP handlers, CLI, migrations). Entries must pass their own field
+// validation, stay within the model's declared rotation length, and never
+// double-book a (week_index, day_of_week) slot.
+func validateModelWithEntries(model *configModels.WorkTimeModel, entries []*configModels.WorkTimeModelEntry) error {
+	if model == nil {
+		return errors.New("model is required")
+	}
+	if err := model.Validate(); err != nil {
+		return err
+	}
+	seenSlots := make(map[string]struct{}, len(entries))
+	for _, entry := range entries {
+		if err := entry.Validate(); err != nil {
+			return err
+		}
+		if entry.WeekIndex >= model.RotationLength {
+			return errors.New("entry week_index outside rotation_length")
+		}
+		slot := fmt.Sprintf("%d:%d", entry.WeekIndex, entry.DayOfWeek)
+		if _, exists := seenSlots[slot]; exists {
+			return errors.New("duplicate entry for week_index and day_of_week")
+		}
+		seenSlots[slot] = struct{}{}
+	}
+	return nil
+}
+
 // UpdateModel replaces the model and its entries, refreshes the schedule
 // snapshots of all staff bound to the template, and returns the reloaded
 // row.
 func (s *WorkTimeModelService) UpdateModel(ctx context.Context, model *configModels.WorkTimeModel, entries []*configModels.WorkTimeModelEntry) (*configModels.WorkTimeModel, error) {
+	if err := validateModelWithEntries(model, entries); err != nil {
+		return nil, err
+	}
 	if err := s.repo.Update(ctx, model, entries); err != nil {
 		return nil, err
 	}
