+++ backend/test/handler_complexity_ratchet_test.go
@@ -40,30 +40,29 @@ var handlerComplexityAllowlist = map[string]int{
 	"api/auth/invitation_handlers.go:(*Resource).createInvitation":        20,
 	"api/base.go:(*API).registerRoutesWithRateLimiting":                   20,
 	"api/base.go:setupCORS": 18,
-	"api/common/student_locations.go:(*StudentLocationSnapshot).ResolveStudentLocationWithTime": 16,
-	"api/enrollment/admin_decision_handlers.go:(*Resource).decideAdminChild":                    20,
-	"api/enrollment/admin_decision_handlers.go:(*Resource).toAdminRequestDetailSummary":         21,
-	"api/enrollment/admin_late_invite_handlers.go:(*Resource).createManualApprovedEnrollment":   20,
-	"api/enrollment/admin_late_invite_handlers.go:(*Resource).getManualEnrollmentBootstrap":     30,
-	"api/enrollment/care_offering_handlers.go:(*Resource).listPublicCareOfferings":              23,
-	"api/enrollment/care_offering_handlers.go:(*Resource).publicFormBootstrap":                  51,
-	"api/enrollment/export_format.go:formatContactList":                                         18,
-	"api/enrollment/legal_handlers.go:(*Resource).publicLegalTexts":                             23,
-	"api/enrollment/phase_handlers.go:(*PhaseRequest).toModel":                                  23,
-	"api/enrollment/report_handlers.go:parseCareUsageFiltersFromQuery":                          16,
-	"api/enrollment/schema_handlers.go:(*Resource).getSchemaPreviewBootstrap":                   20,
-	"api/enrollment/schema_handlers.go:(*Resource).listPublicActiveSchema":                      22,
-	"api/enrollment/schema_handlers.go:(*Resource).publishSchema":                               32,
-	"api/enrollment/schema_handlers.go:(*Resource).updateSchema":                                19,
-	"api/guardians/handlers.go:(*Resource).deleteGuardian":                                      22,
-	"api/iot/checkin/attendance_handlers.go:(*AttendanceResource).handleDailyCheckout":          18,
-	"api/iot/checkin/handlers.go:(*Resource).deviceCheckin":                                     33,
-	"api/iot/checkin/handlers.go:(*Resource).devicePickupQuery":                                 21,
-	"api/iot/checkin/helpers.go:(*Resource).isAfterCheckoutTimeGate":                            17,
-	"api/operator/api.go:(*Resource).Router":                                                    20,
-	"api/operator/settings.go:(*SettingsResource).ResetSchoolSettingValue":                      17,
-	"api/operator/settings.go:(*SettingsResource).SetSchoolSettingValue":                        19,
-	"api/parent/enrollment_handlers.go:(*Resource).submitParentEnrollment":                      20,
+	"api/enrollment/admin_decision_handlers.go:(*Resource).decideAdminChild":                  20,
+	"api/enrollment/admin_decision_handlers.go:(*Resource).toAdminRequestDetailSummary":       21,
+	"api/enrollment/admin_late_invite_handlers.go:(*Resource).createManualApprovedEnrollment": 20,
+	"api/enrollment/admin_late_invite_handlers.go:(*Resource).getManualEnrollmentBootstrap":   30,
+	"api/enrollment/care_offering_handlers.go:(*Resource).listPublicCareOfferings":            23,
+	"api/enrollment/care_offering_handlers.go:(*Resource).publicFormBootstrap":                51,
+	"api/enrollment/export_format.go:formatContactList":                                       18,
+	"api/enrollment/legal_handlers.go:(*Resource).publicLegalTexts":                           23,
+	"api/enrollment/phase_handlers.go:(*PhaseRequest).toModel":                                23,
+	"api/enrollment/report_handlers.go:parseCareUsageFiltersFromQuery":                        16,
+	"api/enrollment/schema_handlers.go:(*Resource).getSchemaPreviewBootstrap":                 20,
+	"api/enrollment/schema_handlers.go:(*Resource).listPublicActiveSchema":                    22,
+	"api/enrollment/schema_handlers.go:(*Resource).publishSchema":                             32,
+	"api/enrollment/schema_handlers.go:(*Resource).updateSchema":                              19,
+	"api/guardians/handlers.go:(*Resource).deleteGuardian":                                    22,
+	"api/iot/checkin/attendance_handlers.go:(*AttendanceResource).handleDailyCheckout":        18,
+	"api/iot/checkin/handlers.go:(*Resource).deviceCheckin":                                   33,
+	"api/iot/checkin/handlers.go:(*Resource).devicePickupQuery":                               21,
+	"api/iot/checkin/helpers.go:(*Resource).isAfterCheckoutTimeGate":                          17,
+	"api/operator/api.go:(*Resource).Router":                                                  20,
+	"api/operator/settings.go:(*SettingsResource).ResetSchoolSettingValue":                    17,
+	"api/operator/settings.go:(*SettingsResource).SetSchoolSettingValue":                      19,
+	"api/parent/enrollment_handlers.go:(*Resource).submitParentEnrollment":                    20,
 	"api/server.go:NewServer":                                                                  35,
 	"api/sse/api.go:(*Resource).eventsHandler":                                                 17,
 	"api/sse/sse_helpers.go:(*Resource).resolveSupervisions":                                   16,
